+++ lib/game_module.rb
@@ -1,15 +1,14 @@
 # frozen_string_literal: true
 
-
 module Enumerable
   def check_diagonal?(side)
     i = -1
     j = 3
-    dia_left = self.all? do |child|
+    dia_left = all? do |child|
       i += 1
       child[i] == side
     end
-    dia_right = self.all? do |child|
+    dia_right = all? do |child|
       j -= 1
       child[j] == side
     end
@@ -18,15 +17,13 @@ module Enumerable
 end
 
 module Error
+  def self.not_available
+    puts 'Position not available, please try again with another position'
+  end
 
-def self.not_available 
-  puts "Position not available, please try again with another position"
-end
-
-def self.not_valid_side
-puts "This is not a valid side, please choose between X and O"
-end
-
+  def self.not_valid_side
+    puts 'This is not a valid side, please choose between X and O'
+  end
 end
 
 #  board class create new board when init
@@ -35,10 +32,10 @@ class Board
   attr_reader :moves, :cells, :has_no_winner
   def initialize
     @cells = [[1, 2, 3], [4, 5, 6], [7, 8, 9]]
-    welcome_note
-    display
     @moves = 9
     @has_no_winner = true
+    welcome_note
+    display
   end
 
   private
@@ -72,10 +69,10 @@ class Board
       @has_no_winner = false
 
     end
-    if (pos % 2 == 1)
+    if pos.odd?
       if @cells.check_diagonal? side
-    puts "Game over, winner is #{side} by dia " 
-      @has_no_winner = false
+        puts "Game over, winner is #{side} by dia "
+        @has_no_winner = false
       end
     end
   end
@@ -88,9 +85,7 @@ class Board
     @moves -= 1
     game_over?(parent_pos, child_pos, side, pos) if @moves <= 5
   end
-
-
-end #end of board class
+end # end of board class
 
 class Player
   attr_accessor :name, :side
@@ -102,70 +97,4 @@ class Player
     puts "Player #{@player_no}: what's your name? "
     @name = gets.chomp
   end
-end #end of player class
-
-board = Board.new
-
-player1 = Player.new(1)
-player2 = Player.new(2)
-player1.set_name
-player2.set_name
-
-
-
-def player1.set_side
-  puts "#{@name} what's your side? Choose between X and O"
-  side = gets.chomp
-  @side = side.upcase
-  puts "Great! #{@name}, your side is #{@side}"
-end
-
-  valid_side = false 
-
-until valid_side  
-
-  player1.set_side
-    
-  if player1.side == "X" || player1.side == "O"
-    valid_side = true 
-  else 
-    Error::not_valid_side
-  end
-end
-
-player2.side = player1.side == 'X' ? 'O' : 'X'
-
-turn = 1 
-
-while board.moves && board.has_no_winner
-  valid_move = false
-  board.display
-
-  until valid_move 
-    puts "#{turn == 1? player1.name : player2.name}, choose me a position"
-    pos = gets.chomp
-    child_pos = (pos.to_f / 3).ceil - 1
-    if (1..9).include?(pos.to_i) && board.cells[child_pos].include?(pos.to_i)
-      valid_move = true
-    else 
-     Error::not_available
-    end
-    
-end 
-
-
-board.update(pos.to_i, turn == 1 ? player1.side: player2.side)
-  
-
-
-#update 
-turn = turn == 1 ?  2 : 1
-
-
-
-  # puts "#{player2.name}, choose a position"
-  # pos = gets.chomp
-  # board.update(pos.to_i, player2.side)
-
-
-end
+end # end of player class
